% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nudz_colors}
\alias{nudz_colors}
\title{Returns HEX color from the NUDZ color list}
\usage{
nudz_colors(...)
}
\arguments{
\item{color}{vector of colors which are translated to their hex interpretations}
}
\value{
hex color
}
\description{
Returns HEX color from the NUDZ color list
}
\examples{
nudz_colors('red')
nudz_colors(c('red', 'light_blue'))
}
