% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_nudz}
\alias{scale_color_nudz}
\title{Scale colors for ggplot color aesthetic}
\usage{
scale_color_nudz(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Scale colors for ggplot color aesthetic
}
