% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_nudz}
\alias{scale_fill_nudz}
\title{Scale colors for ggplot fill aesthetics}
\usage{
scale_fill_nudz(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Scale colors for ggplot fill aesthetics
}
