% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{show_palette}
\alias{show_palette}
\title{Visualizes the palette}
\usage{
show_palette(pal)
}
\arguments{
\item{pal}{created nudz_palette to be visualized}
}
\value{
image of the palette
}
\description{
Visualizes the palette
}
