% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nudz_colors}
\alias{nudz_colors}
\title{Returns HEX color from the NUDZ color list}
\usage{
nudz_colors(...)
}
\arguments{
\item{color}{vector of colors which are translated to their hex interpretations.
Possible values at shi moment are
\itemize{
\item red
\item yellow
\item light_blue
\item dark_blue
\item light_grey
\item dark_grey
\item black
\item white
}}
}
\value{
hex color
}
\description{
Returns HEX color from the NUDZ color list
}
