% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{nudz_colors}
\alias{nudz_colors}
\title{Returns HEX color from the NUDZ color list}
\usage{
nudz_colors(...)
}
\arguments{
\item{color}{}
}
\value{
hex color
}
\description{
Returns HEX color from the NUDZ color list
}
\examples{
colors <- nudz_colors('red')


}
