% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_nudz}
\alias{scale_color_nudz}
\alias{scale_fill_nudz}
\title{Scale colors or fill for ggplot aesthetics}
\usage{
scale_color_nudz(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)

scale_fill_nudz(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{name of the palette to be used. See \code{\link{nudz_palette}}
for all the available options}

\item{discrete}{logical determining if a discreate (finite values) scale
should be created or a continuous interpolation between colors of given palette}

\item{reverse}{shoudl hte order of the colors be reversed}

\item{...}{other parameter passed to \code{\link{ggplot2::discrete_scale}}
or to the \code{\link{ggplot2::scale_color_gradientn}}}
}
\description{
Scale colors or fill for ggplot aesthetics
}
\section{Functions}{
\itemize{
\item \code{scale_fill_nudz}: Scale fill
}}

